Added Tier Color Changing
When selecting a tier, you can change the label and with this PR, change the color. Color values are saved to the json file and will be imported when loading the json file.

Additionally, the tiers are opened by selecting their container, and closed by selecting the container a second time.